applications: refactor helm test to move helper function in dedicated package
the following functions has been moved to pkg/applications/test/helm.go:
  * packageChart
  * StartOciRegistry
  * StartOciRegistryWithAuth
  * newOciRegistry

Code the create http helm registry has been extracted to the following method and moved to pkg/applications/test/helm.go:
  * StartHttpRegistryWithCleanup
  * StartHttpRegistryWithAuthAndCleanup
